refactor(auth): move provider config policy out of agent
Related to #434
