+++ apps/desktop/src/services/AgentService/execution/runtime.ts
@@ -3,6 +3,7 @@
 import { updateModelLastUsed } from '@database/queries';
 import type { SessionTurnEntity } from '@database/types';
 
+import { isTouchAiManagedMode, parseProviderConfigJson } from '@/application/providerConfigPolicy';
 import { t } from '@/i18n';
 import {
     type AttachmentIndex,
@@ -13,7 +14,6 @@ import {
 import type { InputHistorySnapshot } from '@/types/session';
 
 import { AiError, AiErrorCode } from '../contracts/errors';
-import { isTouchAiManagedMode, parseProviderConfigJson } from '../infrastructure/providers/config';
 import { getCurrentModelLanguageContext } from '../languageContext';
 import { PersistenceProjector } from '../outputs/persistence';
 import { composePromptSnapshot } from '../prompt/composer';
